Rename some R&R model objects
Also simplify interface to accept String cluster id.
