+++ src/main/java/com/ibm/watson/developer_cloud/retrieve_and_rank/v1/ClusterLifecycleClient.java
@@ -15,10 +15,9 @@
  */
 package com.ibm.watson.developer_cloud.retrieve_and_rank.v1;
 
+import com.ibm.watson.developer_cloud.retrieve_and_rank.v1.models.SolrClusterList;
+import com.ibm.watson.developer_cloud.retrieve_and_rank.v1.models.SolrClusterOptions;
 import com.ibm.watson.developer_cloud.retrieve_and_rank.v1.models.SolrCluster;
-import com.ibm.watson.developer_cloud.retrieve_and_rank.v1.models.SolrClusterCreationRequest;
-import com.ibm.watson.developer_cloud.retrieve_and_rank.v1.models.SolrClusterListResponse;
-import com.ibm.watson.developer_cloud.retrieve_and_rank.v1.models.SolrClusterResponse;
 
 /**
  * A client for communicating with Watson Search lifecycle APIs.
@@ -27,27 +26,27 @@ public interface ClusterLifecycleClient {
     /**
      * Creates a new Solr cluster.
      */
-    SolrClusterResponse createSolrCluster();
+    SolrCluster createSolrCluster();
 
     /**
      * Creates a new Solr cluster, based on the provided config object.
      */
-    SolrClusterResponse createSolrCluster(SolrClusterCreationRequest creationConfig);
+    SolrCluster createSolrCluster(SolrClusterOptions creationConfig);
 
     /**
      * Deletes an existing Solr Cluster {@code solrCluster}.
      * <p/>
      * Reports success even if no corresponding cluster actually exists.
      */
-    void deleteSolrCluster(SolrCluster solrCluster);
+    void deleteSolrCluster(String solrClusterId);
 
     /**
      * Checks whether a provided Solr cluster is ready for use.
      */
-    SolrClusterResponse pollSolrCluster(SolrCluster solrCluster);
+    SolrCluster pollSolrCluster(String solrClusterId);
 
     /**
      * Lists all Solr clusters currently associated with the service instance.
      */
-    SolrClusterListResponse listSolrClusters();
+    SolrClusterList listSolrClusters();
 }
